Extracted DB handling to the new EventRecorder class (#6)
* Create the EventRecorder class that is responsible for storing events
to the DB.

* Created class to store the application states and changed State values to integers.

* Storing datetime as a TIMESTAMP field.

Later this will return a datetime object when querying statistics

* Adding Unit Tests for the EventRecorder

* Adding Makefile and test requirement files

The test environment can be setup with "make init" and the tests can
be run with "make test".

The requirements on the requirements.test.txt are installed on the
"make init"

* Using "IF NOT EXISTS" to create DB table

Instead of running the query and catching exception.

* Adding docstrings and solving flake8 issues